multiprecision: reorganize files
