+++ crypto3/libs/algebra/include/nil/crypto3/algebra/fields/detail/element/fp.hpp
@@ -31,8 +31,10 @@
 #include <nil/crypto3/algebra/fields/detail/exponentiation.hpp>
 #include <nil/crypto3/algebra/fields/detail/element/operations.hpp>
 
-#include <nil/crypto3/multiprecision/big_uint.hpp>
 #include <nil/crypto3/multiprecision/big_mod.hpp>
+#include <nil/crypto3/multiprecision/big_uint.hpp>
+#include <nil/crypto3/multiprecision/pow.hpp>
+#include <nil/crypto3/multiprecision/ressol.hpp>
 
 #include <type_traits>
