+++ crypto3/libs/algebra/include/nil/crypto3/algebra/fields/detail/element/fp4.hpp
@@ -26,6 +26,8 @@
 #ifndef CRYPTO3_ALGEBRA_FIELDS_ELEMENT_FP4_HPP
 #define CRYPTO3_ALGEBRA_FIELDS_ELEMENT_FP4_HPP
 
+#include <nil/crypto3/multiprecision/wnaf.hpp>
+
 #include <nil/crypto3/algebra/fields/detail/exponentiation.hpp>
 #include <nil/crypto3/algebra/fields/detail/element/operations.hpp>
