+++ crypto3/libs/algebra/include/nil/crypto3/algebra/pairing/mnt6/298/affine_ate_miller_loop.hpp
@@ -26,6 +26,8 @@
 #ifndef CRYPTO3_ALGEBRA_PAIRING_MNT4_298_AFFINE_ATE_MILLER_LOOP_HPP
 #define CRYPTO3_ALGEBRA_PAIRING_MNT4_298_AFFINE_ATE_MILLER_LOOP_HPP
 
+#include <nil/crypto3/multiprecision/wnaf.hpp>
+
 #include <nil/crypto3/algebra/curves/mnt6.hpp>
 #include <nil/crypto3/algebra/pairing/detail/mnt6/298/params.hpp>
 #include <nil/crypto3/algebra/pairing/detail/forms/short_weierstrass/projective/types.hpp>
