+++ crypto3/libs/algebra/include/nil/crypto3/algebra/wnaf.hpp
@@ -29,6 +29,7 @@
 #include <nil/crypto3/algebra/type_traits.hpp>
 
 #include <nil/crypto3/multiprecision/big_uint.hpp>
+#include <nil/crypto3/multiprecision/wnaf.hpp>
 
 namespace nil {
     namespace crypto3 {
