+++ crypto3/libs/multiprecision/include/nil/crypto3/multiprecision/detail/big_int.hpp
@@ -20,7 +20,7 @@
 #include <boost/assert.hpp>
 #include <boost/functional/hash.hpp>
 
-#include "nil/crypto3/multiprecision/detail/big_uint/big_uint_impl.hpp"
+#include "nil/crypto3/multiprecision/big_uint.hpp"
 #include "nil/crypto3/multiprecision/detail/config.hpp"
 
 namespace nil::crypto3::multiprecision {
