--- crypto3/libs/multiprecision/include/nil/crypto3/multiprecision/detail/integer_ops_pow.hpp
@@ -1,11 +1,0 @@
-//---------------------------------------------------------------------------//
-// Copyright (c) 2024 Andrey Nefedov <ioxid@nil.foundation>
-//
-// Distributed under the Boost Software License, Version 1.0
-// See accompanying file LICENSE_1_0.txt or copy at
-// http://www.boost.org/LICENSE_1_0.txt
-//---------------------------------------------------------------------------//
-
-#pragma once
-
-#include "nil/crypto3/multiprecision/detail/big_uint/ops/pow.hpp"  // IWYU pragma: export
