+++ crypto3/libs/multiprecision/include/nil/crypto3/multiprecision/integer.hpp
@@ -9,5 +9,4 @@
 #pragma once
 
 #include "nil/crypto3/multiprecision/detail/integer_ops_base.hpp"  // IWYU pragma: export
-#include "nil/crypto3/multiprecision/detail/integer_ops_pow.hpp"   // IWYU pragma: export
-#include "nil/crypto3/multiprecision/detail/integer_ops_powm.hpp"  // IWYU pragma: export
+#include "nil/crypto3/multiprecision/pow.hpp"                      // IWYU pragma: export
