+++ crypto3/libs/multiprecision/include/nil/crypto3/multiprecision/jacobi.hpp
@@ -10,12 +10,10 @@
 
 #pragma once
 
-// IWYU pragma: private; include "nil/crypto3/multiprecision/big_uint.hpp"
-
 #include <cstddef>
 #include <stdexcept>
 
-#include "nil/crypto3/multiprecision/detail/big_uint/big_uint_impl.hpp"
+#include "nil/crypto3/multiprecision/big_uint.hpp"
 
 namespace nil::crypto3::multiprecision {
 
@@ -23,6 +21,7 @@ namespace nil::crypto3::multiprecision {
     constexpr int jacobi(const big_uint<Bits> &a, const big_uint<Bits> &n) {
         using big_uint_t = big_uint<Bits>;
 
+        // TODO(ioxid): optimize
         if (n % 2u == 0 || n <= 1) {
             throw std::invalid_argument("jacobi: second argument must be odd and > 1");
         }
@@ -40,6 +39,7 @@ namespace nil::crypto3::multiprecision {
                 big_uint_t tmp(y);
                 tmp -= x;
                 x = tmp;
+                // TODO(ioxid): optimize
                 if (y % 4u == 3) {
                     J = -J;
                 }
@@ -51,12 +51,14 @@ namespace nil::crypto3::multiprecision {
             std::size_t shifts = x.lsb();
             x >>= shifts;
             if (shifts & 1) {
+                // TODO(ioxid): optimize
                 std::size_t y_mod_8 = static_cast<std::size_t>(y % 8u);
                 if (y_mod_8 == 3 || y_mod_8 == 5) {
                     J = -J;
                 }
             }
 
+            // TODO(ioxid): optimize
             if (x % 4u == 3 && y % 4u == 3) {
                 J = -J;
             }
