+++ crypto3/libs/multiprecision/test/inverse.cpp
@@ -19,6 +19,7 @@
 
 #include "nil/crypto3/multiprecision/big_mod.hpp"
 #include "nil/crypto3/multiprecision/big_uint.hpp"
+#include "nil/crypto3/multiprecision/inverse.hpp"
 #include "nil/crypto3/multiprecision/literals.hpp"
 
 using namespace nil::crypto3::multiprecision;
